cdk/requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
jsonschema
boto3
boto3-stubs[essential]
pyyaml